[feat]Component 분리
